Add "stats" subcommand and supporting structure
Added a "stats" subcommand to the "cards" console command in main.rs. This new subcommand uses a new "stats" module in cards mod.rs to generate and print statistics about different types of cards. Additionally, refactored the get_card_directory function in cards.rs to be public so it can be used by the new statistics functionality.