Implement feature for card editing before saving
Added a new feature in the card game where users can now edit a card before writing it to a file. The edit functionality is facilitated by the dialoguer's Editor module. If a user confirms to edit, the card's data is fed into an editor where the user can make modifications to it. Deserialize function is added for handling the edited content. All card models are made clonable to support this feature.