Add functionality to draw and shuffle cards
This update establishes the ability to draw and shuffle cards for a game. The new functionality randomly selects cards of different types (event, incident, oopsie and lucky) based on user-specified counts. It also ensures that the deck is properly assembled, incorporating a mechanism to shuffle the cards and then creating a safe and unsafe part of the deck. Lastly, this commit adds a method to acquire all the cards of a specific type.