quick fix for array out of bound panic
when the size of chunks is determined an integer division is performed and the result is truncated. this can lead to more chunks than expected. as the chunks are the elements over which is iterated, iterating over more than evaluation cards are presents this can lead to an array out of bound error with the last, smaller chunk. this fix combines skips the last, smaller chunk if present.