reworks card window and view model
Previously costs where handling oopsie costs and incident_impact. The constructor of the card view model is getting bigger. But as this is basically an anti-corruption-layer between UI and cards this may be okay for the moment. Also, the card window as structured into smaller functions to make clearer how the card windows does look like.