do not reduce attack duration if the attack just started this round
When a new round starts, a card is drawn from the deck. This card is then placed in `board.drawn_card` and `board.open_cards`, reflecting the card to be visible (open) and most recently drawn. This fix omits the most recently drawn card from reducing the duration in case it is an attack card. Before this fix an attack card with e.g., a duration of 2 was drawn, placed into `open_cards` and its duration was reduced directly afterward. This leads to an off by one error.