refactor deck creation to filter by CardCategory
This removes the CardInDeck trait which was simulating some kind of type safetly. However, the `PreparedDeck` struct took only Vectors of `CardRc`. Therefore the illusion of type safety only was a that: An illusion. One could easily have supplied the wrong card type inside such a vec.

It further should be possible to optimize memory more and instead of cloning the Card into the `Deck`  using a `CardRc` there as well. But this is out of scope.